Create Deep_Learning_with_PyTorch_GradCAM.ipynb
Used PyTorch to create and train a CNN Model, hence using the model to implement GradCAM on a simple 3-vegetable dataset. But the model class is created in such a way that it can be generalised for any dataset. This file returns the heat map of a localisation map for a given class, also plotting the heatmap for the given image.